refactor: extract new code into modules to satisfy code-size ratchet
- move rate-limit error formatting into openai_rate_limit_format.rs
- move parse_rate_limit_error into helpers_rate_limit_parse.rs
- move auto-poke error markers into commands_auto_poke_errors.rs
- move placeholder-route helpers + route_supports_reasoning_effort into
  inline_interactive_placeholder_routes.rs
- consolidate subscribe-nudge session fields into SubscribeNudgeState
