+++ crates/jcode-tui/src/tui/app/helpers.rs
@@ -229,97 +229,9 @@ pub(super) fn debug_response_path() -> PathBuf {
     std::env::temp_dir().join("jcode_debug_response")
 }
 
-/// Parse rate limit reset time from error message
-/// Returns the Duration until rate limit resets, if this is a rate limit error
-pub(super) fn parse_rate_limit_error(error: &str) -> Option<Duration> {
-    let error_lower = error.to_lowercase();
-
-    if !error_lower.contains("rate limit")
-        && !error_lower.contains("rate_limit")
-        && !error_lower.contains("429")
-        && !error_lower.contains("too many requests")
-        && !error_lower.contains("hit your limit")
-    {
-        return None;
-    }
-
-    if let Some(idx) = error_lower.find("retry") {
-        let after = &error_lower[idx..];
-        for word in after.split_whitespace() {
-            if let Ok(secs) = word
-                .trim_matches(|c: char| !c.is_ascii_digit())
-                .parse::<u64>()
-                && secs > 0
-                && secs < 86400
-            {
-                return Some(Duration::from_secs(secs));
-            }
-        }
-    }
-
-    if let Some(idx) = error_lower.find("resets") {
-        let after = &error_lower[idx..];
-        for word in after.split_whitespace() {
-            let word = word.trim_matches(|c: char| c == '·' || c == ' ');
-            if (word.ends_with("am") || word.ends_with("pm"))
-                && let Some(duration) = parse_clock_time_to_duration(word)
-            {
-                return Some(duration);
-            }
-        }
-    }
-
-    if let Some(idx) = error_lower.find("reset") {
-        let after = &error_lower[idx..];
-        // Unit-suffixed durations like "resets in 30d 4h 29m" (OpenAI usage
-        // limit messages). Without this, "30d" would parse as 30 seconds and
-        // schedule a bogus 30s auto-retry against a limit that resets in days.
-        let mut unit_total = Duration::ZERO;
-        let mut saw_unit = false;
-        for word in after.split_whitespace().take(8) {
-            let digits: String = word.chars().take_while(|c| c.is_ascii_digit()).collect();
-            let rest = &word[digits.len()..];
-            if digits.is_empty() {
-                continue;
-            }
-            let value: u64 = match digits.parse() {
-                Ok(v) => v,
-                Err(_) => continue,
-            };
-            let secs = match rest.trim_matches(|c: char| !c.is_ascii_alphabetic()) {
-                "d" => Some(value * 86400),
-                "h" => Some(value * 3600),
-                "m" | "min" => Some(value * 60),
-                "s" | "sec" => Some(value),
-                _ => None,
-            };
-            if let Some(secs) = secs {
-                unit_total += Duration::from_secs(secs);
-                saw_unit = true;
-            }
-        }
-        if saw_unit {
-            // Only auto-retry within a day; longer windows should be treated
-            // as terminal by the caller (fallback offer / stop auto-poke).
-            if unit_total > Duration::ZERO && unit_total < Duration::from_secs(86400) {
-                return Some(unit_total);
-            }
-            return None;
-        }
-        for word in after.split_whitespace() {
-            if let Ok(secs) = word
-                .trim_matches(|c: char| !c.is_ascii_digit())
-                .parse::<u64>()
-                && secs > 0
-                && secs < 86400
-            {
-                return Some(Duration::from_secs(secs));
-            }
-        }
-    }
-
-    None
-}
+#[path = "helpers_rate_limit_parse.rs"]
+mod rate_limit_parse;
+pub(super) use rate_limit_parse::parse_rate_limit_error;
 
 pub(super) fn is_context_limit_error(error: &str) -> bool {
     if crate::provider::openai_request::is_openai_encrypted_content_too_large_error(error) {
@@ -1603,33 +1515,4 @@ fn gather_git_info_inner() -> Option<GitInfo> {
         behind,
         dirty_files,
     })
-}
-
-#[cfg(test)]
-mod rate_limit_parse_tests {
-    use super::parse_rate_limit_error;
-    use std::time::Duration;
-
-    #[test]
-    fn usage_limit_reset_in_days_does_not_schedule_bogus_short_retry() {
-        // "30d" must not be misread as 30 seconds.
-        let err = "Rate limited: The usage limit has been reached. Plan: team. \
-                   Resets in 30d 4h 29m (2026-08-21 04:31 UTC).";
-        assert_eq!(parse_rate_limit_error(err), None);
-    }
-
-    #[test]
-    fn unit_suffixed_reset_within_a_day_is_parsed() {
-        let err = "429 rate limit exceeded. Resets in 2h 5m.";
-        assert_eq!(
-            parse_rate_limit_error(err),
-            Some(Duration::from_secs(2 * 3600 + 5 * 60))
-        );
-    }
-
-    #[test]
-    fn plain_retry_seconds_still_parse() {
-        let err = "429 Too Many Requests: retry after 30 seconds";
-        assert_eq!(parse_rate_limit_error(err), Some(Duration::from_secs(30)));
-    }
 }
